requirements.txt
openai
langchain
langchain-community
langchain-openai
sentence-transformers
faiss-cpu
python-dotenv
flask